Made pseudonyms stateful and reworked pseudonym editing interface
`Assassin` now also stores a dict `pseudonym_datetimes` that gives the start of validiity for each pseudonym. If a pseudonym does not appear in the dict then it is assumed to be always valid.

I reworked the pseudonym editing interface so that pseudonyms can be edited without deleting them (which would mess up indices). Deleting a pseudonym is replaced by setting a pseudonym to blank.

The new interface also asks the umpire to specify the start of validity for each pseudonym they add. This defaults to the current datetime (which is good enough for most cases, but sometimes the umpire may need to change it). The initial pseudonym cannot be deleted or have a validity set -- it is always valid forever.