[63] Assassin -> Resurrect as Police (#69)
The interface asks the user to choose from all the dead non-police assassins, then asks for a new pseudonym for the "police clone" of this assassin.
Assassin objects now have a `clone` method that returns a copy of the assassin with a new identifier  and _secret_id, and specified changes. In particular, Assassin -> Resurrect as Police  sets the clone to police and its pseudonyms to the single pseudonym entered.
The original assassin is hidden after cloning to avoid confusion.