[17] Created ScoringPlugin (#72)
Created ScoringPlugin

This plugin generates the Open Season page. It exposes config exports allowing a start datetime to be set for open season, and to set the scoring formula.

Scoring is implemented using a ScoreManager which processes events in a similar way to DeathManager etc. From the events it constructs the kill tree and counts the numbers of attempts made by each player, and tracks which players are still alive (and hence should appear on the open season page).

As a stopgap, `eval` is used to parse the scoring formula, which can be set by the user.