Implemented `DangerousConfigExport` (#83)
To do this, the way that the `options_functions` attribute of plugin exports is handled was modified to allow any HTMLcomponent to be displayed. A special `ConfigOptionsList` component is used to render DangerousConfigExports appropriately -- for inquirer_cli this is done by colouring them red using ANSI codes. The component also challenges attempts to access the config, using the same method as the `Reset Database` export of `LocalBackupPlugin`.

`colorama` is used in the inquirer cli to ensure the colours work on Windows, but if that module is missing the error is suppressed.