[102] [SRCFPlugin] Auto-backups and remote backup sorting (#109)
* Autobackups

`SRCF -> Upload database and [...]` now creates a remote backup at the same time as uploading the database. This backup is based on the user's *local* database, not the remote database. The auto-generated name includes the username of the user who caused the backup to be created.

* Custom sorting of remote backups

This sorting extracts the date and time from the backup names in order to sort them in reverse-chronological order.