[116] Prevent creation of non-Police assassins after first email (#121)
* Prevented non-Police assassins being created after first email sent

This is to prevent the targeting graph being messed up. Conversely, I have now allowed player types to be changed before the first email is sent.

* `force_default` attribute for `Checkbox`

This is now used to force players to prevent non-police being created after the game start.

* Apply suggestions from code review

Co-authored-by: J. Syiek <43001551+jsyiek@users.noreply.github.com>

* Response to (remaining) comments

Moved police Checkbox back to original place (now that it has interaction), and changed the formulation of the condition to detect whether the game has started.

* Removed rogue `print`

---------

Co-authored-by: J. Syiek <43001551+jsyiek@users.noreply.github.com>