More useful targeting info in AssassinDependentCrimeEntry
The component now takes the targeting graph as a parameter, rather than generic licitness info, to allow it to filter to actually relevant targeting info, because the other licitness info proved to be problematic to implement in the current framework.

Fetching the targeting graph is done through a new 'data hook' framework.