[46] Include targeting info of victims in AssassinDependentCrimeEntry (#154)
* Include targeting info of victims in AssassinDependentCrimeEntry

Want to also include wantedness, inco status, but this would require event datetime to be a dependency....

* More useful targeting info in AssassinDependentCrimeEntry

The component now takes the targeting graph as a parameter, rather than generic licitness info, to allow it to filter to actually relevant targeting info, because the other licitness info proved to be problematic to implement in the current framework.

Fetching the targeting graph is done through a new 'data hook' framework.

* bugfix: don't display targeting info when TargetingPlugin disabled

Before, when TargetingPlugin was disabled it would claim any kill might be illicit.