[140] `Setup Game` export (#163)
* Created `Setup Game` export

* Added game start to Setup Game

Included some help text in the config option, and fixed a bug where CorePlugin has its on_setup_game called twice.

* Added competency config to Setup Game

* Added police config to Setup Game

Only selection of Umpire and CoP since these are what actually matter.

* Require players to be added before using Setup Game

* Added targeting config to Setup Game

* Added UIConfig to Setup Game

* Removed May Week setup bcs incomplete

* Allow `Setup Game` to enable/disable plugins even if no players added yet

This is so that it can ensure that `PlayerImporterPlugin` is enabled in the case that an umpire runs `Setup Game` on a fresh install.

* Added abstract HiddenComponent parent class for HiddenTextbox and HiddenJSON

* Response to (rest of) code review